refactor: 상수분리 , 변수명 변경
