+++ index.html
@@ -8,7 +8,7 @@
     <title>Simple Notion App</title>
   </head>
   <body>
-    <main class="app"></main>
+    <div class="app"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
