+++ src/Constant/error.js
@@ -1,0 +1,9 @@
+export const ERROR_MESSAGE = {
+  NOT_STRING: '문자열이 아닙니다.',
+  NOT_NUMBER: '숫자가 아닙니다.',
+  NOT_DOCUMENTS_ROOT: '올바른 주소가 아닙니다',
+  NOT_CONSTRUCTOR: '생성자 함수가 아닙니다.',
+  NOT_VALID_LOCALDATA: '로컬 데이터 로드에 에러가 발생했습니다',
+  NOT_COLLECT_NETWORK: '네트워크 응답이 올바르지 않습니다.',
+  NOT_COLLECT_ADRESS: '올바르지 않은 주소입니다.',
+};
