+++ src/Constant/normal.js
@@ -1,0 +1,8 @@
+export const MINIMUM_INPUT_LENGTH = 2;
+export const CHANGE_API_DATA_TO_LOCAL_DATA =
+  '이전 저장된 임시 글이 있습니다. 업로드 하시겠습니까? 저장되지 않은 글은 삭제됩니다.';
+export const ALERT_DELETE_DOCUMENT = '글을 삭제하시겠습니까? 하위 폴더는 자동으로 이동됩니다.';
+export const CHANGE_USER_NAME = '변경할 USER ID를 입력해 주세요';
+export const NEW_TITLE = '타이틀을 입력해 주세요';
+export const NEW_CONTENT = '내용을 입력해 주세요';
+export const NEW_ROOT_ID = 'RootDocument';
