+++ src/Helpers/getUserIdToAdress.js
@@ -1,5 +1,3 @@
-import { isString } from './checkError.js';
-
 export const getUserIdToAdress = () => {
   const { pathname } = location;
   const [, , userId] = pathname.split('/');
